place tab related code in correct directories
