+++ DSNP-Wallet/Source/Common/TabBar/Permissions Screen/PermissionsView.swift
@@ -1,0 +1,26 @@
+//
+//  PermissionsView.swift
+//  DSNP-Wallet
+//
+//  Created by Ben Gabay on 6/1/23.
+//
+
+import SwiftUI
+
+struct PermissionsView: View {
+    
+    var body: some View {
+        VStack {
+            
+        }
+        .frame(maxWidth: .infinity, maxHeight: .infinity)
+        .ignoresSafeArea()
+        .background(Color(uiColor: UIColor.Theme.bgTeal))
+    }
+}
+
+struct PermissionsView_Previews: PreviewProvider {
+    static var previews: some View {
+        PermissionsView()
+    }
+}
