+++ DSNP-Wallet/Source/Common/TabBar/Profile Screen/ProfileView.swift
@@ -1,0 +1,27 @@
+//
+//  ProfileView.swift
+//  DSNP-Wallet
+//
+//  Created by Ben Gabay on 6/1/23.
+//
+
+import SwiftUI
+
+struct ProfileView: View {
+    
+    var body: some View {
+        VStack {
+            
+        }
+        .frame(maxWidth: .infinity, maxHeight: .infinity)
+        .ignoresSafeArea()
+        .background(Color(uiColor: UIColor.Theme.bgTeal))
+    }
+}
+
+
+struct ProfileView_Previews: PreviewProvider {
+    static var previews: some View {
+        ProfileView()
+    }
+}
